Bulk: Fix threshold handling.
The bulk could have consumed more messages than you have configured, e.
g. five messages of type LobbyEntry for the same player would be
merged into one message and thus not reaching a threshold of three.

Furthermore some message types require to flush the bulk whether or not
it reached the threshold, for example the LobbyRemove-Message.